Version: 1.4.4 | Testing 1. Changed the folder name from giveaways 2. Now completely rewritten to the new OOP 3. Removed unnecessary comments and in particular the code 4. Other fixes
