+++ src/main/java/startbot/BotStart.java
@@ -2,8 +2,8 @@ package startbot;
 
 import config.Config;
 import events.MessageWhenBotJoinToGuild;
-import giftaway.MessageGift;
-import giftaway.Reactions;
+import giveaway.MessageGift;
+import giveaway.Reactions;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -20,10 +20,9 @@ import net.dv8tion.jda.api.entities.Activity;
 
 public class BotStart {
 
-  public static JDA jda;
+  private static JDA jda;
   private final JDABuilder jdaBuilder = JDABuilder.createDefault(Config.getTOKEN());
   private static final Map<String, String> mapPrefix = new HashMap<>();
-//  private static final Map<String, String> idMessagesWithGiveawayEmoji = new HashMap<>();
 
   public void startBot() throws Exception {
     jdaBuilder.setAutoReconnect(true);
@@ -44,20 +43,10 @@ public class BotStart {
       Statement statement = conn.createStatement();
       String sql = "select * from prefixs";
       ResultSet rs = statement.executeQuery(sql);
-//      Statement statementSecond = conn.createStatement();
-
-//      String sqlIdMessagesWithPollEmoji = "select * from idMessagesWithGiveawayEmoji";
-//      ResultSet rsIdMessages = statementSecond.executeQuery(sqlIdMessagesWithPollEmoji);
 
       while (rs.next()) {
         mapPrefix.put(rs.getString("serverId"), rs.getString("prefix"));
       }
-
-//      while (rsIdMessages.next()) {
-//        idMessagesWithGiveawayEmoji.put(
-//            rsIdMessages.getString("idMessagesWithGiveawayEmoji"),
-//            rsIdMessages.getString("idMessagesWithGiveawayEmoji"));
-//      }
 
     } catch (SQLException e) {
       e.printStackTrace();
@@ -68,11 +57,7 @@ public class BotStart {
     return mapPrefix;
   }
 
-//  public static Map<String, String> getIdMessagesWithGiveawayEmoji() {
-//    return idMessagesWithGiveawayEmoji;
-//  }
-
-  public JDA getJda() {
+  public static JDA getJda() {
     return jda;
   }
 }
